Build root/server.js with Webpack
Prior to this commit, we were restricted on how we could write
root/server.js and any files it imports:

 * No Flow syntax (other than comment syntax)
 * No JSX
 * No ES modules
 * No ProvidePlugin imports

Because root/server.js was executed directly in Node.js, all files had
to conform with normal JavaScript rules and syntax.  (This is why
certain files have had comments saying not to use these features.)

This commit builds root/server.js with Webpack, which means our server
code can now be written without restrictions.  The issue with this is
how to reload the server when things change.

 * In development, we want the server to reload whenever changes to
   components or the server itself are made.  For this, we now use
   nodemon to watch root/static/build/server.js for changes and restart
   the server in its entirety.

 * In production, we want the server to reload *gracefully* when DBDefs
   changes and SIGHUP is received.  For this, we use the start_server
   superdaemon from Server::Starter, which is what we already use for
   plackup in production.  It works perfectly fine for Node.js too by
   specifying the file descriptor to use in an environment variable.

I still think it's useful to be able to run root/server.js directly
without a build step (at least for development), but this is true of
other modules too.  While this commit may seem like a step backwards in
that respect, it's not an unachievable goal:

 * Recent versions of Node.js can execute ES modules directly without
   any experimental flag.

 * ./bin/sucrase-node can be used to execute ES modules that use Flow or
   JSX syntax.

As a last step, I think we should consider removing the ProvidePlugin,
because it ties our code too closely to Webpack.  It does make using
certain utilities more convenient, but Flow's IDE integration (related
to autoimports) has also recently improved to where this isn't as much
of a time-saver as before.

+++ root/server.js
@@ -15,10 +15,10 @@ const fs = require('fs');
 const spawnSync = require('child_process').spawnSync;
 
 const Sentry = require('@sentry/node');
-const yargs = require('yargs');
 
 const createServer = require('./server/createServer');
 const DBDefs = require('./static/scripts/common/DBDefs');
+const nonEmpty = require('./static/scripts/common/utility/nonEmpty').default;
 const writeCoverage = require('./utility/writeCoverage');
 
 function sentryInit(config) {
@@ -30,24 +30,32 @@ function sentryInit(config) {
 }
 sentryInit(DBDefs);
 
-yargs
-  .option('socket', {
-    alias: 's',
-    default: DBDefs.RENDERER_SOCKET,
-    describe: 'UNIX socket path',
-  })
-  .option('workers', {
-    alias: 'w',
-    default: process.env.RENDERER_WORKERS || 1,
-    describe: 'Number of workers to spawn',
-  });
-
-const SOCKET_PATH = yargs.argv.socket;
-const WORKER_COUNT = yargs.argv.workers;
+const WORKER_COUNT = parseInt(process.env.RENDERER_WORKERS, 10) || 1;
 const DISCONNECT_TIMEOUT = 10000;
+const SERVER_STARTER_PORT = process.env.SERVER_STARTER_PORT;
+
+let SERVER_STARTER_FD = null;
+let SOCKET_PATH = null;
+
+if (nonEmpty(SERVER_STARTER_PORT)) {
+  const fdMatch = SERVER_STARTER_PORT.match(/=([0-9]+)$/);
+  const fd = fdMatch ? parseInt(fdMatch[1], 10) : NaN;
+  if (Number.isNaN(fd)) {
+    throw new Error(
+      'Invalid file descriptor in SERVER_STARTER_PORT: ' +
+      JSON.stringify(SERVER_STARTER_PORT),
+    );
+  }
+  SERVER_STARTER_FD = fd;
+} else {
+  SOCKET_PATH = process.env.RENDERER_SOCKET;
+  if (!nonEmpty(SOCKET_PATH)) {
+    SOCKET_PATH = DBDefs.RENDERER_SOCKET;
+  }
+}
 
 if (cluster.isMaster) {
-  if (fs.existsSync(SOCKET_PATH)) {
+  if (SOCKET_PATH != null && fs.existsSync(SOCKET_PATH)) {
     if (spawnSync('lsof', [SOCKET_PATH]).status) {
       fs.unlinkSync(SOCKET_PATH);
     } else {
@@ -98,13 +106,6 @@ if (cluster.isMaster) {
     }
   }
 
-  function disconnectWorker(worker) {
-    if (worker.isConnected()) {
-      worker.disconnect();
-      setTimeout(() => killWorker(worker), DISCONNECT_TIMEOUT);
-    }
-  }
-
   function cleanup() {
     const timeout = setTimeout(() => {
       for (const id in cluster.workers) {
@@ -119,50 +120,16 @@ if (cluster.isMaster) {
     });
   }
 
-  let hupAction = null;
-  function hup() {
-    console.info('master received SIGHUP; restarting workers');
-
-    Sentry.close().then(function () {
-      sentryInit(require('./static/scripts/common/DBDefs'));
-    });
-
-    let oldWorkers;
-    let initialTimeout = 0;
-
-    if (hupAction) {
-      clearTimeout(hupAction);
-      initialTimeout = 2000;
-    }
-
-    function killNext() {
-      if (!oldWorkers) {
-        oldWorkers = Object.values(cluster.workers);
-      }
-      const oldWorker = oldWorkers.pop();
-      if (oldWorker) {
-        const doKill = function () {
-          disconnectWorker(oldWorker);
-          hupAction = setTimeout(killNext, 1000);
-        };
-        if (!forkWorker(doKill)) {
-          doKill();
-        }
-      } else {
-        hupAction = null;
-      }
-    }
-
-    hupAction = setTimeout(killNext, initialTimeout);
-  }
-
   process.on('SIGINT', cleanup);
   process.on('SIGTERM', cleanup);
-  process.on('SIGHUP', hup);
 } else {
-  createServer(SOCKET_PATH);
+  if (SERVER_STARTER_FD == null) {
+    createServer(SOCKET_PATH);
+  } else {
+    createServer({fd: SERVER_STARTER_FD});
+  }
 
-  process.on('beforeExit', function () {
+  cluster.worker.on('disconnect', function () {
     const coverage = global.__coverage__;
     if (coverage) {
       writeCoverage(
